+++ lib/strategies/index.js
@@ -1,3 +1,3 @@
 module.exports = {
-    macd : require('./macd.old')
+    macd : require('./macd')
 }
